loss-order -- замена создания FXMLLoader на один fxmlManager -- добавлен priceStep (буден необходим для поднятия шага цены)

+++ src/ru/ex4/apibt/logic/Wait.java
@@ -45,12 +45,12 @@ public class Wait {
     }
 
 
-    public static void sleep(long minutes, String msg) {
+    public static void sleep(double minutes, String msg) {
         try {
             if (msg != null) {
                 Logs.info(msg);
             }
-            Thread.sleep(1000 * 60 * minutes);
+            Thread.sleep((long)(1000 * 60 * minutes));
         } catch (InterruptedException ex) {
             Thread.currentThread().interrupt();
         }
